chore: refactor KVArchiverDataStore to pull out KVContractDataStore
