+++ yarn-project/archiver/src/archiver/index.ts
@@ -3,4 +3,6 @@ export * from './config.js';
 export { type PublishedL2Block, type L1PublishedData } from './structs/published.js';
 export type { ArchiverDataStore } from './archiver_store.js';
 export { KVArchiverDataStore, ARCHIVER_DB_VERSION } from './kv_archiver_store/kv_archiver_store.js';
+export type { ContractDataStore } from './kv_archiver_store/contract_data_store.js';
+export { KVContractDataStore } from './kv_archiver_store/kv_contract_data_store.js';
 export { ContractInstanceStore } from './kv_archiver_store/contract_instance_store.js';
