+++ yarn-project/archiver/src/archiver/kv_archiver_store/kv_contract_data_store.test.ts
@@ -1,0 +1,15 @@
+import { openTmpStore } from '@aztec/kv-store/lmdb-v2';
+
+import { describeContractDataStore } from './contract_data_store_test_suite.js';
+import { KVContractDataStore } from './kv_contract_data_store.js';
+
+describe('KVContractDataStore', () => {
+  let contractStore: KVContractDataStore;
+
+  beforeEach(async () => {
+    const store = await openTmpStore('contract_data_store_test');
+    contractStore = new KVContractDataStore(store);
+  });
+
+  describeContractDataStore('ContractDataStore', () => contractStore);
+});
