+++ yarn-project/end-to-end/src/devnet/e2e_smoke.test.ts
@@ -15,8 +15,7 @@ import { FeeJuiceContract } from '@aztec/noir-contracts.js/FeeJuice';
 import { TestContract } from '@aztec/noir-test-contracts.js/Test';
 import type { AztecNode } from '@aztec/stdlib/interfaces/client';
 import { deriveSigningKey } from '@aztec/stdlib/keys';
-import { registerInitialSandboxAccountsInWallet } from '@aztec/test-wallet/server';
-import { TestWallet } from '@aztec/test-wallet/server';
+import { TestWallet, registerInitialSandboxAccountsInWallet } from '@aztec/test-wallet/server';
 
 import { exec } from 'node:child_process';
 import { lookup } from 'node:dns/promises';
